修复显示费用总额过大报错

+++ src/main/java/com/lotut/pms/dao/FeeDao.java
@@ -57,9 +57,9 @@ public interface FeeDao {
 	
 	List<Fee> getFeesForPatentId(long patentId);
 	
-	int getFeeAllCountByUser(int userId);
+	long getFeeAllCountByUser(int userId);
 	
-	int getUnPaidCountByUser(int userId);
+	long getUnPaidCountByUser(int userId);
 
 	List<String> getFeeTypes(String appNo);
